Fix: 로그인 셋업/테스트 파일 병합 (#168)
* Test: 로그인 정보 저장/ 로그인 확인 테스트

* Test: 로그인 테스트 코드 작성

* Refactor: 콘솔제거

* Test: 멘토링 리뷰 반영

* Test: 로그인 셋업/테스트 파일 병합

* Comment: 테스트명수정

* Refactor: 로그인페이지로 이동 코드 위치 수정, 경로 수정
